Refactor root tests to use root client
The tests themselves are unchanged.
